Renamed Notification component to Alert and vice versa and updated dependencies
